backend/foodgram_back/requirements.txt
Django==3.2
gunicorn==20.1.0
Pillow==9.0.1
djangorestframework==3.12.4
pytz==2022.7
psycopg2-binary==2.8.6
djoser==2.1.0
python-dotenv==0.21.1
djangorestframework-simplejwt==4.7.2
pytest==7.2.2
pytest-django==4.5.2
django-base64field==1.0
django-cors-headers==3.14.0